Move helper method to implementation. Change auth response to actual type. Update unit tests.
