+++ src/main/java/dev/codesupport/web/common/service/controller/AuthenticationController.java
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
 import java.io.Serializable;
-import java.util.List;
 
 /**
  * Controller interface for defining mappings and annotations.
@@ -23,20 +22,9 @@ import java.util.List;
 public interface AuthenticationController {
 
     @PostMapping(value = "/authenticate")
-    ResponseEntity<RestResponse<Serializable>> authenticate(@RequestBody @Valid AuthenticationRequest authenticationRequest);
+    ResponseEntity<RestResponse<String>> authenticate(@RequestBody @Valid AuthenticationRequest authenticationRequest);
 
     @GetMapping(value = "/authenticate/discord")
     ResponseEntity<RestResponse<Serializable>> linkDiscord(@RequestParam String code);
-
-    /**
-     * Returns a new instance of {@link RestResponse}
-     * <p>This exists to make unit testing easier</p>
-     *
-     * @param objectList The resources to include in the {@link RestResponse}
-     * @return The expected {@link RestResponse}
-     */
-    default RestResponse<Serializable> getRestResponse(List<Serializable> objectList) {
-        return new RestResponse<>(objectList);
-    }
 
 }
